requirements.txt
annotated-types==0.7.0
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
fake-useragent==2.0.3
future==1.0.0
google-auth==2.37.0
google-genai==0.5.0
Google-Search-API==1.1.14
idna==3.10
importlib_resources==6.5.2
multidict==6.1.0
pillow==11.1.0
propcache==0.2.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.5
pydantic_core==2.27.2
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
selenium==2.53.6
soupsieve==2.6
typing_extensions==4.12.2
Unidecode==1.3.8
urllib3==1.26.20
vcrpy==7.0.0
websockets==14.2
wrapt==1.17.2
yarl==1.18.3
zipp==3.21.0
